Adaptive layout implementation.
